STAC-23143: move fake instanceconfig to common module
