+++ splunk_base/tests/common.py
@@ -2,6 +2,7 @@
 # All rights reserved
 # Licensed under a 3-clause BSD style license (see LICENSE)
 from stackstate_checks.dev import get_docker_hostname
+from stackstate_checks.splunk.config import AuthType
 
 HOST = get_docker_hostname()
 PORT = '8089'
@@ -59,3 +60,24 @@ def empty_instance_jwt(initial_token):
         'saved_searches': [],
         'collection_interval': 15
     }
+
+
+class FakeInstanceConfig(object):
+    def __init__(self):
+        self.base_url = 'http://testhost:8089'
+        self.default_request_timeout_seconds = 10
+        self.verify_ssl_certificate = False
+        self.ignore_saved_search_errors = True
+        self.username = "admin"
+        self.audience = "test"
+        self.name = "admin"
+        self.token_expiration_days = 90
+        self.renewal_days = 10
+        self.initial_token = "asdfg"
+        self.auth_type = AuthType.BasicAuth
+        self.keyfile = ""
+        self.cert = ""
+        self.timeout = 5000
+
+    def get_auth_tuple(self):
+        return ('username', 'password')
